refactor: navigation, mapper, data 패키지 core로 이동 및 dummy 제거 #1
