--- app/src/main/java/com/ganaljigi/kubf/dummy/data/repository/DummyRepository.kt
@@ -1,7 +1,0 @@
-package com.ganaljigi.kubf.dummy.data.repository
-
-import com.ganaljigi.kubf.dummy.data.response.DummyResponseDto
-
-interface DummyRepository {
-    suspend fun dummy(): Result<DummyResponseDto>
-}
