--- app/src/main/java/com/ganaljigi/kubf/dummy/data/response/DummyResponseDto.kt
@@ -1,10 +1,0 @@
-package com.ganaljigi.kubf.dummy.data.response
-
-import kotlinx.serialization.SerialName
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class DummyResponseDto(
-    @SerialName("description")
-    val description: String
-)
