--- app/src/main/java/com/ganaljigi/kubf/dummy/data/service/DummyService.kt
@@ -1,9 +1,0 @@
-package com.ganaljigi.kubf.dummy.data.service
-
-import com.ganaljigi.kubf.dummy.data.response.DummyResponseDto
-import retrofit2.http.GET
-
-interface DummyService {
-    @GET("/dummy-api")
-    suspend fun getDummy(): DummyResponseDto
-}
