--- app/src/main/java/com/ganaljigi/kubf/dummy/ui/dummy/model/DummyUiState.kt
@@ -1,5 +1,0 @@
-package com.ganaljigi.kubf.dummy.ui.dummy.model
-
-data class DummyUiState(
-    val description: String = "",
-)
